% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_gfonts}
\alias{check_gfonts}
\title{Checks the operability of 'gfonts'}
\usage{
check_gfonts(errors = FALSE)
}
\arguments{
\item{errors}{if TRUE, function is triggering errors if a condition
is not satisfied.}
}
\value{
TRUE or FALSE
}
\description{
Checks that 'gfonts' is installed and can be used.
Packages 'curl' and 'gfonts' must be installed and internet
connectivity must be active.
}
\keyword{internal}
