% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-caching.R
\name{dummy_setup}
\alias{dummy_setup}
\title{Set up a temporary font cache}
\usage{
dummy_setup()
}
\value{
The path to the temporary cache directory (invisibly).
}
\description{
Create a temporary directory used as the
'Google Fonts' cache. This is intended for examples and tests.
}
\keyword{internal}
