% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_set.R
\name{font_google}
\alias{font_google}
\title{Google Font specification}
\usage{
font_google(family, subset = c("latin", "latin-ext"))
}
\arguments{
\item{family}{family name of a 'Google Fonts', for example,
"Open Sans", "Roboto", "Fira Code" or "Fira Sans Condensed".}

\item{subset}{font subset, a character vector, it defaults to
only "latin" and "latin-ext".}
}
\value{
An object of class \code{font_spec}.
}
\description{
Create a font specification for a Google Font,
to be used with \code{\link[=font_set]{font_set()}}.
}
\examples{
font_google("Roboto")
}
\seealso{
Other font set functions:
\code{\link[=font_liberation]{font_liberation()}},
\code{\link[=font_set]{font_set()}},
\code{\link[=font_set_auto]{font_set_auto()}},
\code{\link[=font_set_liberation]{font_set_liberation()}}
}
\concept{font set functions}
