% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/font_metrics.R
\name{font_metrics}
\alias{font_metrics}
\title{Get font metrics for a string of text}
\usage{
font_metrics(str, bold = FALSE, italic = FALSE, fontname = "Sans",
  fontsize = 12)
}
\arguments{
\item{str}{string of text}

\item{bold}{is bold}

\item{italic}{is italic}

\item{fontname}{font name}

\item{fontsize}{font size in pixel}
}
\description{
Calculate the font metrics for a string of text as
a numeric named vector of length 3 containing width, ascent
and descent.
}
\examples{
font_metrics(str = "Hello World!", bold = TRUE, italic = FALSE,
  fontname = "Sans", fontsize = 12)
}

