% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontconfig.R
\name{fontconfig_reinit}
\alias{fontconfig_reinit}
\title{reload Fontconfig configuration}
\usage{
fontconfig_reinit()
}
\description{
This function can be used to make fontconfig
reload font configuration files.
}
\author{
Paul Murrell
}
