% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-proxies.R
\name{glyphs_match}
\alias{glyphs_match}
\title{Validate glyph entries}
\usage{
glyphs_match(x, fontname = "sans", bold = FALSE, italic = FALSE,
  fontfile = "")
}
\arguments{
\item{x}{Character vector of strings}

\item{fontname}{Font name}

\item{bold, italic}{Is text bold/italic?}

\item{fontfile}{Font file}
}
\value{
a logical vector, if a character element is containing at least
a glyph that can not be matched in the font table, FALSE is returned.
}
\description{
Determines if strings contain glyphs not part of a font.
}
\examples{
glyphs_match(letters)
glyphs_match("\\u265E", bold = TRUE)
}
