% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font-caching.R
\name{has_internet}
\alias{has_internet}
\title{Test for internet connectivity}
\usage{
has_internet()
}
\description{
The function is a copy of
\code{\link[curl:nslookup]{curl::has_internet()}}.
}
\examples{
has_internet()
}
\keyword{internal}
