% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{installed_gfonts}
\alias{installed_gfonts}
\title{List installed 'Google Fonts'}
\usage{
installed_gfonts()
}
\value{
families names as a character vector
}
\description{
List installed 'Google Fonts' that can be
found in the user cache directory.
}
\examples{
\dontrun{
if (check_gfonts()) {
  dummy_setup()
  register_gfont(family = "Roboto")
  installed_gfonts()
}
}
}
\seealso{
Other functions for font management:
\code{\link[=addGFontHtmlDependency]{addGFontHtmlDependency()}},
\code{\link[=fonts_cache_dir]{fonts_cache_dir()}},
\code{\link[=gfontHtmlDependency]{gfontHtmlDependency()}},
\code{\link[=install_gfont_script]{install_gfont_script()}},
\code{\link[=liberationmonoHtmlDependency]{liberationmonoHtmlDependency()}},
\code{\link[=liberationsansHtmlDependency]{liberationsansHtmlDependency()}},
\code{\link[=liberationserifHtmlDependency]{liberationserifHtmlDependency()}},
\code{\link[=register_gfont]{register_gfont()}},
\code{\link[=register_liberationmono]{register_liberationmono()}},
\code{\link[=register_liberationsans]{register_liberationsans()}},
\code{\link[=register_liberationserif]{register_liberationserif()}}
}
\concept{functions for font management}
