% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberation.R
\name{liberationmonoHtmlDependency}
\alias{liberationmonoHtmlDependency}
\title{'Liberation Mono' Font HTML dependency}
\usage{
liberationmonoHtmlDependency()
}
\description{
Create an HTML dependency ready
to be used in 'Shiny' or 'R Markdown' with
'Liberation Mono' Font.
}
\seealso{
\code{\link[=font_set_liberation]{font_set_liberation()}}, \code{\link[=font_set]{font_set()}}, \code{\link[=gfontHtmlDependency]{gfontHtmlDependency()}}

Other functions for font management:
\code{\link[=addGFontHtmlDependency]{addGFontHtmlDependency()}},
\code{\link[=fonts_cache_dir]{fonts_cache_dir()}},
\code{\link[=gfontHtmlDependency]{gfontHtmlDependency()}},
\code{\link[=install_gfont_script]{install_gfont_script()}},
\code{\link[=installed_gfonts]{installed_gfonts()}},
\code{\link[=liberationsansHtmlDependency]{liberationsansHtmlDependency()}},
\code{\link[=liberationserifHtmlDependency]{liberationserifHtmlDependency()}},
\code{\link[=register_gfont]{register_gfont()}},
\code{\link[=register_liberationmono]{register_liberationmono()}},
\code{\link[=register_liberationsans]{register_liberationsans()}},
\code{\link[=register_liberationserif]{register_liberationserif()}}
}
\concept{functions for font management}
