% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{png_as_base64}
\alias{png_as_base64}
\title{Encode a png file into base64.}
\usage{
png_as_base64(filename)
}
\arguments{
\item{filename}{name of the file to create}
}
\description{
Encode a png file into base64.
}
\examples{
r <- as.raster(matrix(hcl(50, 80, seq(50, 80, 10)),
 nrow = 4, ncol = 5))

t <- tempfile()
raster_write(r, t, width = 50, height = 50)
png_as_base64(t)
}

