% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberation.R
\name{register_liberationsans}
\alias{register_liberationsans}
\title{Register font 'Liberation Sans'}
\usage{
register_liberationsans(name = "Liberation Sans")
}
\arguments{
\item{name}{the name to use for the font family when
registering with 'systemfonts'. Using a custom name
(e.g. \code{"sans"}, \code{"serif"}, \code{"mono"}) allows devices like
'ragg' to resolve generic family names to this font.}
}
\value{
TRUE if the operation went ok.
}
\description{
Register font 'Liberation Sans' so that it can be used
with devices using the 'systemfonts' package, i.e. the 'flextable'
package and graphic outputs generated with the 'ragg', 'svglite'
and 'ggiraph' packages.
}
\examples{
register_liberationsans()
register_liberationsans(name = "sans")
}
\seealso{
\code{\link[=font_set_liberation]{font_set_liberation()}}, \code{\link[=font_set]{font_set()}}

Other functions for font management:
\code{\link[=addGFontHtmlDependency]{addGFontHtmlDependency()}},
\code{\link[=fonts_cache_dir]{fonts_cache_dir()}},
\code{\link[=gfontHtmlDependency]{gfontHtmlDependency()}},
\code{\link[=install_gfont_script]{install_gfont_script()}},
\code{\link[=installed_gfonts]{installed_gfonts()}},
\code{\link[=liberationmonoHtmlDependency]{liberationmonoHtmlDependency()}},
\code{\link[=liberationsansHtmlDependency]{liberationsansHtmlDependency()}},
\code{\link[=liberationserifHtmlDependency]{liberationserifHtmlDependency()}},
\code{\link[=register_gfont]{register_gfont()}},
\code{\link[=register_liberationmono]{register_liberationmono()}},
\code{\link[=register_liberationserif]{register_liberationserif()}}
}
\concept{functions for font management}
