% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontconfig.R
\name{set_dummy_conf}
\alias{set_dummy_conf}
\alias{unset_dummy_conf}
\title{Set and unset a minimalistic Fontconfig configuration}
\usage{
set_dummy_conf()

unset_dummy_conf()
}
\description{
Set and unset a minimalistic Fontconfig configuration
}
\examples{
if( require("fontquiver") ){
on_appveyor <- function() {
  identical(Sys.getenv("APPVEYOR"), "True")
}

# Use minimal fonts.conf to speed up fc-cache
if (on_appveyor()) {
  set_dummy_conf()
}

unset_dummy_conf()
}
}
