% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_extents}
\alias{str_extents}
\title{Compute string extents.}
\usage{
str_extents(str, bold = FALSE, italic = FALSE, fontname = "Sans",
  fontsize = 12L)
}
\arguments{
\item{str}{Character ector of of strings to measure}

\item{bold,italic}{Is text bold/italic?}

\item{fontname}{Font name}

\item{fontsize}{Font size (in pixels)}
}
\description{
Determines the width and height of a bounding box that's big enough
to (just) enclose the provided text.
}
\examples{
str_extents(letters)
str_extents("Hello World!", bold = TRUE, italic = FALSE,
  fontname = "Sans", fontsize = 12)
}

