% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_family_exists.R
\name{sys_fonts}
\alias{sys_fonts}
\title{List fonts for 'systemfonts'.}
\usage{
sys_fonts()
}
\value{
A data.frame of font information.
}
\description{
List system and registry fonts into a data.frame
containing columns such as path, family, style, weight and italic.
}
\examples{
sys_fonts()
}
