% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/gdtools.R
\name{version_freetype}
\alias{version_freetype}
\alias{version_cairo}
\title{Version numbers of C libraries}
\usage{
version_freetype()

version_cairo()
}
\value{
An object of class \code{"numeric_version"}.
}
\description{
Return the runtime version of the Cairo and FreeType
libraries linked to the package.
}
\examples{
version_cairo()
version_freetype()
}
