% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_raster.R
\name{write_raster}
\alias{write_raster}
\title{Draw a raster to a png file}
\usage{
write_raster(x, filename, width = 480, height = 480, interpolate = FALSE)
}
\arguments{
\item{x}{A raster object}

\item{filename}{name of the file to create}

\item{width,height}{Width and height in pixels.}

\item{interpolate}{A logical value indicating whether to linearly
interpolate the image.}
}
\description{
Draw a raster to a png file
}
\examples{
r <- as.raster(matrix(hcl(0, 80, seq(50, 80, 10)),
 nrow = 4, ncol = 5))
plot(r)
t <- tempfile()
write_raster(r, t, width = 50, height = 50)
}

