% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_set.R
\name{font_liberation}
\alias{font_liberation}
\title{Liberation Font specification}
\usage{
font_liberation(variant = c("sans", "serif", "mono"))
}
\arguments{
\item{variant}{one of \code{"sans"}, \code{"serif"}, or \code{"mono"}.}
}
\value{
An object of class \code{font_spec}.
}
\description{
Create a font specification for a Liberation font,
to be used with \code{\link[=font_set]{font_set()}}.
}
\examples{
font_liberation("sans")
font_liberation("mono")
}
\seealso{
Other font set functions: 
\code{\link{font_google}()},
\code{\link{font_set}()},
\code{\link{font_set_auto}()},
\code{\link{font_set_liberation}()}
}
\concept{font set functions}
