% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_set.R
\name{font_set_liberation}
\alias{font_set_liberation}
\title{Liberation font set}
\usage{
font_set_liberation()
}
\value{
An object of class \code{font_set}.
}
\description{
Shortcut to create a \code{\link[=font_set]{font_set()}} with all four
font roles using Liberation fonts (Sans, Serif, Mono) and
Liberation Sans as a fallback for symbols.
}
\examples{
fonts <- font_set_liberation()
fonts
}
\seealso{
Other font set functions: 
\code{\link{font_google}()},
\code{\link{font_liberation}()},
\code{\link{font_set}()},
\code{\link{font_set_auto}()}
}
\concept{font set functions}
