% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{installed_gfonts}
\alias{installed_gfonts}
\title{List installed 'Google Fonts'}
\usage{
installed_gfonts()
}
\value{
families names as a character vector
}
\description{
List installed 'Google Fonts' that can be
found in the user cache directory.
}
\examples{
\dontrun{
if (check_gfonts()) {
  dummy_setup()
  register_gfont(family = "Roboto")
  installed_gfonts()
}
}
}
\seealso{
Other functions for font management: 
\code{\link{addGFontHtmlDependency}()},
\code{\link{fonts_cache_dir}()},
\code{\link{gfontHtmlDependency}()},
\code{\link{install_gfont_script}()},
\code{\link{liberationsansHtmlDependency}()},
\code{\link{register_gfont}()},
\code{\link{register_liberationsans}()}
}
\concept{functions for font management}
