% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{liberationsansHtmlDependency}
\alias{liberationsansHtmlDependency}
\title{'Liberation Sans' Font HTML dependency}
\usage{
liberationsansHtmlDependency()
}
\description{
Create an HTML dependency ready
to be used in 'Shiny' or 'R Markdown' with
'Liberation Sans' Font.
}
\seealso{
\code{\link[=gfontHtmlDependency]{gfontHtmlDependency()}}

Other functions for font management: 
\code{\link{addGFontHtmlDependency}()},
\code{\link{fonts_cache_dir}()},
\code{\link{gfontHtmlDependency}()},
\code{\link{install_gfont_script}()},
\code{\link{register_gfont}()},
\code{\link{register_liberationsans}()}
}
\concept{functions for font management}
