% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberation.R
\name{liberationserifHtmlDependency}
\alias{liberationserifHtmlDependency}
\title{'Liberation Serif' Font HTML dependency}
\usage{
liberationserifHtmlDependency()
}
\description{
Create an HTML dependency ready
to be used in 'Shiny' or 'R Markdown' with
'Liberation Serif' Font.
}
\seealso{
\code{\link[=font_set_liberation]{font_set_liberation()}}, \code{\link[=font_set]{font_set()}}, \code{\link[=gfontHtmlDependency]{gfontHtmlDependency()}}

Other functions for font management: 
\code{\link{addGFontHtmlDependency}()},
\code{\link{fonts_cache_dir}()},
\code{\link{gfontHtmlDependency}()},
\code{\link{install_gfont_script}()},
\code{\link{installed_gfonts}()},
\code{\link{liberationmonoHtmlDependency}()},
\code{\link{liberationsansHtmlDependency}()},
\code{\link{register_gfont}()},
\code{\link{register_liberationmono}()},
\code{\link{register_liberationsans}()},
\code{\link{register_liberationserif}()}
}
\concept{functions for font management}
