% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{match_family}
\alias{match_family}
\title{Find best family match with fontconfig}
\usage{
match_family(family = "sans", bold = 1L, italic = 1L)
}
\arguments{
\item{family}{Family to match.}

\item{bold}{Wheter to match a font featuring a \code{bold} face.}

\item{italic}{Wheter to match a font featuring an \code{italic} face.}
}
\description{
This returns the best font family match for the pattern
constructed with \code{bold} and \code{italic}. The default
pattern is bold italic to make sure the matched font has enough
features to be used in R graphics (plain, bold, italic, bold
italic).
}
\examples{
match_family("sans")
match_family("serif", bold = FALSE, italic = TRUE)
}

