% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberation.R
\name{register_liberationmono}
\alias{register_liberationmono}
\title{Register font 'Liberation Mono'}
\usage{
register_liberationmono()
}
\value{
TRUE if the operation went ok.
}
\description{
Register font 'Liberation Mono' so that it can be used
with devices using the 'systemfonts' package, i.e. the 'flextable'
package and graphic outputs generated with the 'ragg', 'svglite'
and 'ggiraph' packages.
}
\seealso{
Other functions for font management: 
\code{\link{addGFontHtmlDependency}()},
\code{\link{fonts_cache_dir}()},
\code{\link{gfontHtmlDependency}()},
\code{\link{install_gfont_script}()},
\code{\link{installed_gfonts}()},
\code{\link{liberationmonoHtmlDependency}()},
\code{\link{liberationsansHtmlDependency}()},
\code{\link{liberationserifHtmlDependency}()},
\code{\link{register_gfont}()},
\code{\link{register_liberationsans}()},
\code{\link{register_liberationserif}()}
}
\concept{functions for font management}
