% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liberation-sans.R
\name{register_liberationsans}
\alias{register_liberationsans}
\title{Register font 'Liberation Sans'}
\usage{
register_liberationsans()
}
\value{
TRUE if the operation went ok.
}
\description{
Register font 'Liberation Sans' so that it can be used
with devices using the 'systemfonts' package, i.e. the 'flextable'
package and graphic outputs generated with the 'ragg', 'svglite'
and 'ggiraph' packages.
}
\seealso{
Other functions for font management: 
\code{\link{addGFontHtmlDependency}()},
\code{\link{fonts_cache_dir}()},
\code{\link{gfontHtmlDependency}()},
\code{\link{install_gfont_script}()},
\code{\link{installed_gfonts}()},
\code{\link{liberationsansHtmlDependency}()},
\code{\link{register_gfont}()}
}
\concept{functions for font management}
