% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/str_extents.R
\name{str_extents}
\alias{str_extents}
\title{get extents for strings of text}
\usage{
str_extents(str, bold = FALSE, italic = FALSE, fontname = "Sans",
  fontsize = 12)
}
\arguments{
\item{str}{strings of text}

\item{bold}{is bold}

\item{italic}{is italic}

\item{fontname}{font name}

\item{fontsize}{font size in pixel}
}
\description{
Gets the extents for strings of text. The extents describe a
rectangle that encloses the portion of the formatted text.
}
\examples{
str_extents(str = "Hello World!", bold = TRUE, italic = FALSE,
  fontname = "Sans", fontsize = 12)
}

