% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_metrics}
\alias{str_metrics}
\title{Get font metrics for a string.}
\usage{
str_metrics(str, bold = FALSE, italic = FALSE, fontname = "sans",
  fontsize = 12L)
}
\arguments{
\item{str}{string of text}

\item{bold}{Is text bold/italic?}

\item{italic}{Is text bold/italic?}

\item{fontname}{Font name}

\item{fontsize}{Font size (in pixels)}
}
\value{
A named numeric vector
}
\description{
Get font metrics for a string.
}
\examples{
str_metrics("Hello World!")
}

