% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{str_metrics}
\alias{str_metrics}
\title{Get font metrics for a string.}
\usage{
str_metrics(x, fontname = "sans", fontsize = 12L, bold = FALSE,
  italic = FALSE)
}
\arguments{
\item{x}{Character vector of strings to measure}

\item{fontname}{Font name}

\item{fontsize}{Font size (in pixels)}

\item{bold}{Is text bold/italic?}

\item{italic}{Is text bold/italic?}

\item{str}{string of text}
}
\value{
A named numeric vector
}
\description{
Get font metrics for a string.
}
\examples{
str_metrics("Hello World!")
}

