% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp-proxies.R
\name{str_metrics}
\alias{str_metrics}
\title{Get font metrics for a string.}
\usage{
str_metrics(x, fontname = "sans", fontsize = 12, bold = FALSE,
  italic = FALSE, fontfile = "")
}
\arguments{
\item{x}{Character vector of strings to measure}

\item{fontname}{Font name}

\item{fontsize}{Font size}

\item{bold}{Is text bold/italic?}

\item{italic}{Is text bold/italic?}

\item{fontfile}{Font file}
}
\value{
A named numeric vector
}
\description{
Get font metrics for a string.
}
\examples{
\donttest{
# The first run can be slow when font caches are missing
# as font files are then being scanned to build those font caches.
str_metrics("Hello World!")
}
}
