% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sys_fonts}
\alias{sys_fonts}
\title{List system fonts.}
\usage{
sys_fonts()
}
\description{
List system fonts details into a data.frame containing columns foundry, family,
file, slant and weight.
}
\examples{
\donttest{
# The first run can be slow when font caches are missing
# as font files are then being scanned to build those font caches.
sys_fonts()
}
}
