% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_family_exists.R
\name{sys_fonts}
\alias{sys_fonts}
\title{List fonts for 'systemfonts'.}
\usage{
sys_fonts()
}
\description{
List system and registryfonts details into a data.frame
containing columns foundry, family, file, slant and weight.
}
\examples{
sys_fonts()
}
