% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_raster.R
\name{write_raster}
\alias{write_raster}
\title{draw a raster image to a png file}
\usage{
write_raster(red, green, blue, alpha, w, h, width, height,
  interpolate = FALSE, filename)
}
\arguments{
\item{red}{integer value for red coumpounds (between 0 and 255).}

\item{green}{integer value for green coumpounds (between 0 and 255).}

\item{blue}{integer value for blue coumpounds (between 0 and 255).}

\item{alpha}{integer value for alpha coumpounds (between 0 and 255).}

\item{w}{columns number of the image data.}

\item{h}{rows number of the image data.}

\item{width}{width of the image in pixels.}

\item{height}{height of the image in pixels.}

\item{interpolate}{A logical value indicating whether to linearly
interpolate the image.}

\item{filename}{name of the file to create}
}
\description{
Draw a raster image based on colors. Image data is arranged by
row, it's made of 4 elements giving one colour (red, green, blue and alpha).
Data is scaled to width and height.
}
\examples{
# create a color data matrix
w <- 5L
h <- 4L
red <- c(197L, 197L, 197L, 197L, 197L, 225L, 225L, 225L, 225L,
225L, 254L, 254L, 254L, 254L, 254L, 255L, 255L, 255L, 255L, 255L)
green = c(78L, 78L, 78L, 78L, 78L, 106L, 106L, 106L, 106L, 106L,
134L, 134L, 134L, 134L, 134L, 162L, 162L, 162L, 162L, 162L)
blue = c(109L, 109L, 109L, 109L, 109L, 134L, 134L, 134L, 134L,
134L, 161L, 161L, 161L, 161L, 161L, 188L, 188L, 188L, 188L, 188L)
alpha = rep(255L, length(blue))

# generate the corresponding png
write_raster(red = as.integer(red), green = as.integer(green),
blue = as.integer(blue), alpha = as.integer(alpha),
width = 50, height = 75, w = w, h = h, filename = "raster.png")
}

